changing direct function name
